Change scripts for saving TensorRT model and importing native codes
